Tranformando um uma util medidorForcaSenha
